Add support for artificial slicing.
Artificial slicing is when a play's meta data is missing or
corrupt, so that the timing for it cannot be retrieved. Thus,
the play footage cannot be sliced. Instead, a static image with
the play description is turned into a 10 second video clip
to compensate. (The idea is to provide the presence of absence.)